pyproject.toml
[tool.black]
line-length = 88
target-version = ['py38']
include = '\.py$'
exclude = '''
(
  /(
      \.
    | \.git
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.pytest_cache
    | \.idea
    | \.cache
    | bin
    | config
    | build
    | dist
  )/
)
'''

[tool.poetry]
name = "stop-the-vcr"
version = "0.1.2"
description = "A package that contains VCR.py custom request matchers"
authors = ["Alex Cornelio <ascornelio@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.8"
vcrpy = "^4.2.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "^22.12.0"
flake8 = "^5.0.0"  # TODO: since flake8 ^6.0.0 requires python > 3.8.1, update flake8 when deprecating python 3.8
pre-commit = "^2.21.0"
mypy = "^0.991"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
